+++ src/main/java/com/melog/melog/emotion/adapter/emotion/EmotionKeywordJpaRepository.java
@@ -1,7 +1,7 @@
 package com.melog.melog.adapter.out.persistence.emotion;
 
-import com.melog.melog.domain.emotion.EmotionKeyword;
-import com.melog.melog.domain.emotion.EmotionRecord;
+import com.melog.melog.emotion.domain.EmotionKeyword;
+import com.melog.melog.emotion.domain.EmotionRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
