+++ src/main/java/com/melog/melog/emotion/adapter/emotion/EmotionScorePersistenceAdapter.java
@@ -1,9 +1,9 @@
 package com.melog.melog.adapter.out.persistence.emotion;
 
 import com.melog.melog.application.port.out.EmotionScorePersistencePort;
-import com.melog.melog.domain.emotion.EmotionRecord;
-import com.melog.melog.domain.emotion.EmotionScore;
-import com.melog.melog.domain.emotion.EmotionType;
+import com.melog.melog.emotion.domain.EmotionRecord;
+import com.melog.melog.emotion.domain.EmotionScore;
+import com.melog.melog.emotion.domain.EmotionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
